+++ projects/miopen/src/solver/conv/conv_asm_implicit_gemm_gtc_bwd_nhwc.cpp
@@ -766,7 +766,7 @@ bool PerformanceConfigAsmImplicitGemmGTCBwdXdlopsNHWC::IsValidValue() const
     if(IsDefaultConstructed())
         return true;
     const auto& config_list = GetBwdXdlopsNHWCConfigList();
-    if(index < config_list.size() && *this == config_list[index])
+    if(static_cast<size_t>(index) < config_list.size() && *this == config_list[index])
         return true;
     return miopen::any_of(config_list, [&](auto v) { return (*this == v); });
 }
@@ -793,7 +793,7 @@ bool PerformanceConfigAsmImplicitGemmGTCBwdXdlopsNHWC::SetNextValue(const Proble
             {
                 index++;
             }
-            if(index >= config_list.size())
+            if(static_cast<size_t>(index) >= config_list.size())
                 return false;
             CopyParameters(config_list[index]);
         }
