+++ projects/miopen/src/solver/conv/conv_asm_implicit_gemm_gtc_fwd_nchwc.cpp
@@ -455,7 +455,7 @@ bool PerformanceConfigAsmImplicitGemmGTCFwdDlopsNCHWC::SetNextValue(const Proble
         else
         {
             index++;
-            if(index >= config_list.size())
+            if(static_cast<size_t>(index) >= config_list.size())
                 return false;
             CopyParameters(config_list[index]);
         }
@@ -473,7 +473,7 @@ bool PerformanceConfigAsmImplicitGemmGTCFwdDlopsNCHWC::IsValidValue() const
     if(IsDefaultConstructed())
         return true;
     const auto& config_list = GetFwdDlopsNCHWCConfigList();
-    if(index < config_list.size() && *this == config_list[index])
+    if(static_cast<size_t>(index) < config_list.size() && *this == config_list[index])
         return true;
     return miopen::any_of(config_list, [&](auto v) { return (*this == v); });
 }
