+++ projects/miopen/src/solver/conv/conv_asm_implicit_gemm_gtc_wrw_nhwc.cpp
@@ -732,7 +732,7 @@ bool PerformanceConfigAsmImplicitGemmGTCWrwXdlopsNHWC::SetNextValue(const Proble
             {
                 index++;
             }
-            if(index >= config_list.size())
+            if(static_cast<size_t>(index) >= config_list.size())
                 return false;
             CopyParameters(config_list[index]);
         }
@@ -750,7 +750,7 @@ bool PerformanceConfigAsmImplicitGemmGTCWrwXdlopsNHWC::IsValidValue() const
     if(IsDefaultConstructed())
         return true;
     const auto& config_list = GetWrwXdlopsNHWCConfigList();
-    if(index < config_list.size() && *this == config_list[index])
+    if(static_cast<size_t>(index) < config_list.size() && *this == config_list[index])
         return true;
     return miopen::any_of(config_list, [&](auto v) { return (*this == v); });
 }
