+++ projects/miopen/src/solver/conv/conv_hip_implicit_gemm_bwd_v4r1.cpp
@@ -833,7 +833,8 @@ ConvHipImplicitGemmBwdDataV4R1::GetSolution(const ExecutionContext& ctx,
     assert(config.IsValid(problem));
 
     // a series of kernels
-    for(std::size_t gemm_id = 0; gemm_id < CalculateNumberOfGemm(problem); ++gemm_id)
+    for(std::size_t gemm_id = 0; gemm_id < static_cast<size_t>(CalculateNumberOfGemm(problem));
+        ++gemm_id)
     {
         KernelInfo construction_parameters;
